Consolidate primitive ID lowering
